fixing task 10

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- Create the SafeDiv function
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 BEGIN
     IF b = 0 THEN
         RETURN 0; -- Return 0 if the second number is 0
